Change UX

+++ app.py
@@ -1,11 +1,12 @@
 from urllib import request
-from flask import Flask,render_template,url_for,request,redirect
+from flask import Flask, render_template, url_for, request, redirect
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
 db = SQLAlchemy(app)
+
 
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
@@ -14,6 +15,7 @@ class Todo(db.Model):
 
     def __repr__(self):
         return '<Task %r>' % self.id
+
 
 @app.route('/', methods=['POST', 'GET'])
 def index():
@@ -33,23 +35,28 @@ def index():
 
         return render_template('index.html', tasks=tasks)
 
-@app.route('/delete/<int:id>')
+
+@app.route('/delete/<int:id>', methods=['GET', 'POST'])
 def delete(id):
     task_to_delete = Todo.query.get_or_404(id)
 
-    try:
-        db.session.delete(task_to_delete)
-        db.session.commit()
-        return redirect('/')
-    except:
-        return "There was an error deleting the task"
+    if request.method == 'POST':
+        try:
+            db.session.delete(task_to_delete)
+            db.session.commit()
+            return redirect('/')
+        except:
+            return "There was an error deleting the task"
+    else:
+        return render_template('delete.html', task=task_to_delete)
+
 
 @app.route('/update/<int:id>', methods=['GET', 'POST'])
 def update(id):
 
     task = Todo.query.get_or_404(id)
     if request.method == 'POST':
-        
+
         task.content = request.form['content']
 
         try:
@@ -62,4 +69,4 @@ def update(id):
 
 
 if __name__ == "__main__":
-    app.run(debug = True) 
+    app.run(debug=True)
